requirements.txt
Flask==2.2.5
keras==3.0.5
keras==2.10.0
matplotlib==3.8.4
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
